Add files via upload
In this updated version, the script uses the PowerShell command Start-Process with the -Verb RunAs parameter to request administrator privileges. It will relaunch the script with elevated permissions. If the UAC prompt appears, you can provide the necessary credentials to run the script as an administrator.

Please note that the Optimize-Volume PowerShell command assumes the drive letter is "C". If you want to optimize a different drive, modify the -DriveLetter parameter accordingly.